+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.81] - 2024-04-20
+### Added
+* some `EditorBuildSettingsScene` functions to `Utilities.Editor.General`
+### Changed
+* `ScenesInBuild` file making the script easier to read
+* `ScenesInBuild` auto-generated file making its reserialize and retain its scene object better (not perfect though)
+
 ## [0.0.80] - 2024-04-19
 ### Added
 * `EditorOnly` `PropertyDrawer` to `ScenesInBuildManager`'s auto-generated class